+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Socket enable/disable option for *Shade Object Smooth* node.
 - Support undo for node operators.
 - Increased default spline resolution for Spline nodes.
+- Inprove spiral in Distribute Matrices node.
 
 ## 2.2.0 (01 September 2020)